Adding viewer methods to work show presenter decorator
