+++ app/presenters/hyku/work_show_presenter_decorator.rb
@@ -15,6 +15,22 @@ module Hyku
       klass == JournalArticle ? 'Periodical' : 'Part Of'
     end
 
+    def pdf_viewer?
+      return unless Flipflop.default_pdf_viewer?
+      return unless file_set_presenters.any?(&:pdf?) || pdf_extension?
+
+      # If all of the member_presenters are file_set presenters, return true
+      # this also means that there are no child works
+      member_presenters.all? { |presenter| presenter.is_a? Hyrax::FileSetPresenter }
+    end
+
+    def pdf_extension?
+      file_set_presenters.any? { |fsp| fsp.label.downcase.end_with?('.pdf') }
+    end
+
+    def viewer?
+      iiif_viewer? || video_embed_viewer? || pdf_viewer?
+    end
     private
 
       def model_klass
